When a polynomial is created in `polycal`, it is done using a finite range of independent values, and domain of dependent values. This is the only data for which the fit can be assumed to be accurate: no data was provided for predictions to be made outside this range.
The previous version of `polycal` returned an error in the forward
problem if the argument was outside the range of independent values used
to create the polynomial, and one in the backward direction if the
argument was outside the range of dependent values used to create the
polynomial. No information was attached to the error, other than the
fact the input was out of bounds.

In some applications like numerical optimisation it is useful to query
the objective function continuously across the parameter space. Not
being able to do this prevented users from safely using, for example,
Gauss Newton optimisation.

To resolve this new error variants were added to describe the case where
the input is out-of-bounds. An error is still returned
(`OutOfBoundsCertain`, `OutOfBoundsUncertain`), but the error now
contains the value requested.

This means that users can choose to access and use the result of their
request if necessary.

Note that monotonicity is only checked on the *original* range of the
function. It is entirely possible roots may be repeated outside this
range. If the polynomial is queried out of range, then the caller should
always verify the result of the optimisation is within the original
range of the function.