Corrected Weighted Least Squares Implementation
In a previous release the definition of uncertainties in the constructor
and polynomial sections of the code had decoupled from that in the
underlying implementation of weighted least squares. Variances passed in
from the interface code were treated as standard uncertainties. This
resulted in the code reporting associated errors which deviated from
their true values by around a factor of a square root. This change corrects this:

* Implementations of weighted and total least squares now understand the
  uncertainties provided are variances. The struct `Uncertainty` in the
  `solver` submodule was renamed to `Covariance` to make it clear what
  the contained values signify.